pyproject.toml
[tool.poetry]
name = "watchcat"
version = "0.1.0"
description = "A transparent overlay window application for desktop automation"
authors = ["mark <mark@cs-magic.com>"]
packages = [{include = "src"}]
license = "MIT"
repository = "https://github.com/cs-magic-open/watchcat"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
PyQt6 = "^6.4.0"
opencv-python = "^4.8.0"
numpy = "^1.26.0"
mss = "^9.0.1"
plyer = "^2.1.0"
notify-py = "^0.3.43"
pydub = "^0.25.1"
pydantic = "^2.10.3"
mkdocs-video = "^1.5.0"
cairosvg = "^2.7.1"

[tool.poetry.group.linux.dependencies]
simpleaudio = { version = "^1.0.4", platform = "linux" }

[tool.poetry.group.windows.dependencies]
sounddevice = { version = "^0.4.6", platform = "win32" }

[tool.poetry.group.macos.dependencies]
pyobjc-framework-cocoa = { version = "^10.3.2", platform = "darwin" }
pyobjc-framework-quartz = { version = "^10.3.2", platform = "darwin" }
pyobjc-framework-applicationservices = { version = "^10.3.2", platform = "darwin" }
pyobjc-framework-coretext = { version = "^10.3.2", platform = "darwin" }
sounddevice = { version = "^0.4.6", platform = "darwin" }
pyobjus = { version = "^1.2.1", platform = "darwin" }
rumps = { version = "^0.4.0", platform = "darwin" }

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-qt = "^4.2.0"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.5.0"
mkdocs = "^1.5.0"
pymdown-extensions = "^10.5"

[tool.poetry.group.dev.dependencies]
briefcase = "^0.3.20"
py2app = { version = "^0.28.8", platform = "darwin" }
pyinstaller = "^6.11.1"
cairosvg = "^2.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.briefcase]
project_name = "WatchCat"
bundle = "com.cs-magic.watchcat"
version = "0.1.0"
url = "https://github.com/cs-magic-open/watchcat"
author = "mark"
author_email = "mark@cs-magic.com"
license = "MIT"

[tool.briefcase.app.src]
formal_name = "WatchCat"
description = "A transparent overlay window application for desktop automation"
icon = "resources/icon"
sources = ["src"]
requires = []

[tool.briefcase.app.src.macOS]
requires = []
universal_build = true
supported = true
minimum_os_version = "10.15"